Update tests with synth ReOrg
