+++ fooof/tests/test_synth_gen.py
@@ -1,70 +1,13 @@
-"""Test functions for FOOOF synth."""
-
-from collections import Iterable
-from itertools import repeat
+"""Test functions for FOOOF synth.gen"""
 
 import numpy as np
 
 from fooof.tests.utils import default_group_params
 
-from fooof.synth import *
-from fooof.synth import _check_flat, _check_iter
+from fooof.synth.gen import *
 
 ###################################################################################################
 ###################################################################################################
-
-def test_param_iter():
-
-    # Test oscillations
-    step = Stepper(8, 12, .1)
-    osc = [step, .5, .5]
-    iter_1 = param_iter(osc)
-
-    for ind, val in enumerate(iter_1):
-        assert val == [8 + (.1*ind), .5 , .5]
-
-    # Test background
-    step = Stepper(.25, 3, .25)
-    bg = [0, step]
-    iter_1 = param_iter(bg)
-
-    for ind, val in enumerate(iter_1):
-        assert val == [0, .25 + (.25*ind)]
-
-    # Test n oscillations
-    step = Stepper(8, 12, .1)
-    oscs = [step, .5, .5, 10, .25, 1]
-    iter_1 = param_iter(oscs)
-
-    for ind, val in enumerate(iter_1):
-        assert val == [8 + (.1*ind), .5 , .5, 10, .25, 1]
-
-    # Test list of lists
-    step = Stepper(8, 12, .1)
-    osc_1 = [1, 2, 3]
-    osc_2 = [4, 5, 6]
-    osc_3 = [7, 8, step]
-    oscs = [osc_1, osc_2, osc_3]
-    iter_2 = param_iter(oscs)
-
-    for ind, val in enumerate(iter_2):
-        assert val == [1, 2, 3, 4, 5, 6, 7, 8, 8 + (.1*ind)]
-
-
-def test_stepper():
-
-    assert Stepper(8,12,.1)
-
-    # TODO: add more tests of Stepper
-
-
-def test_param_sampler():
-
-    pos = [1, 2, 3, 4]
-    gen = param_sampler(pos)
-
-    for ind, el in zip(range(3), gen):
-        assert el in pos
 
 def test_gen_freqs():
 
@@ -148,54 +91,3 @@ def test_gen_power_values():
     ys = gen_power_vals(xs, bg_params, gauss_params, nlv)
 
     assert np.all(ys)
-
-def test_rotate_spectrum():
-
-    # Create a spectrum to use for test rotations
-    freqs, spectrum = gen_power_spectrum([1, 100], [1, 1], [])
-
-    # Check that rotation transforms the power spectrum
-    rotated_spectrum = rotate_spectrum(freqs, spectrum, delta_f=0.5, f_rotation=25.)
-    assert not np.all(rotated_spectrum == spectrum)
-
-    # Check that 0 rotation returns the same spectrum
-    rotated_spectrum = rotate_spectrum(freqs, spectrum, delta_f=0., f_rotation=25.)
-    assert np.all(rotated_spectrum == spectrum)
-
-
-def test_check_iter():
-
-    # Note: generator case not tested
-
-    # Check that a number input becomes an iterable
-    out = _check_iter(12, 3)
-    assert isinstance(out, Iterable)
-    assert isinstance(out, repeat)
-
-    # Check that single list becomes repeat iterable
-    out = _check_iter([1, 1], 2)
-    assert isinstance(out, Iterable)
-    assert isinstance(out, repeat)
-
-    # Check that a list of lists, of right length stays list of list
-    out = _check_iter([[1, 1], [1, 1], [1, 1]], 3)
-    assert isinstance(out, Iterable)
-    assert isinstance(out, list)
-    assert isinstance(out[0], list)
-
-def test_check_flat():
-
-    # Check an empty list stays the same
-    assert _check_flat([]) == []
-
-    # Check an already flat list gets left the same
-    lst = [1, 2, 3, 4]
-    flat_lst = _check_flat(lst)
-    assert flat_lst == lst
-
-    # Check a nested list gets flattened
-    lst = [[1, 2], [3, 4]]
-    flat_lst = _check_flat(lst)
-    for el in flat_lst:
-        assert isinstance(el, int)
-    assert len(flat_lst) == 4
